Revert "Adds a workaround to fix version inconsistency"
This reverts commit e2e1683720bb54d7d662573204fd2a5b7cc6e4c6.

Reason for revert: It was much to my chagrin that I realized the main purpose of this build target is an actually an internal branch which has the changes which make S the default for Robolectric.  This will complicate things a little bit in the short term, but in principle, this revert can be merged over with those changes down the line.

Change-Id: I053a32d89bfdf33486d95a30b145eb3b6f57df1d